fastapi/requirements.txt
boto3==1.42.9
fastapi==0.124.4
mangum==0.19.0
mypy-boto3-dynamodb==1.42.3
PyJWT==2.10.1
